Added quite a few new features. Now you can change from hex to text view with an option in the command, or a shortcut key inside the program itself.
You are now able to run commands inside FAT, by pressing `O`.

There are some other enhancements and a new doc.

Readme.md has to be updated, and I caught a heisenbug with the search option (when there's no valid coincidence it causes a buffer overflow in release, while in debug there's no issue. Magic behind the compiler).

These two things will be the next objective, but tomorrow. Now it's too late. I will open an issue now as a reminder (and to see if someone is willing to do the dirty job) for the heisenbug.